move to arch_params_factory
