+++ src/super_gradients/training/models/__init__.py
@@ -19,4 +19,5 @@ from super_gradients.training.models.segmentation_models.shelfnet import *
 from super_gradients.training.models.classification_models.efficientnet import *
 from super_gradients.training.models.all_architectures import ARCHITECTURES, ModelNames
 from super_gradients.training.models.user_models import *
-from super_gradients.training.models.model_factory import get, get_arch_params
+from super_gradients.training.models.model_factory import get
+from super_gradients.training.models.arch_params_factory import get_arch_params
